requirements.txt
selenium==4.2.0
bs4
dateparser==1.1.8
six
kafka-python
pymysql
lxml
certifi==2022.12.7
charset-normalizer==2.1.1
gevent==22.10.2
greenlet==2.0.1
idna==3.4
requests==2.28.1
urllib3==1.26.13
websocket-client==1.4.2
zope.event==4.6
zope.interface==5.5.2
unidecode==1.3.6